Brenehams algorithm
Testing a variation of Breneham's algorithm to find the grids affected by a line - this will be important for any line-based abilities in the future, and will probably be extensible to cover other shapes (cone = lines drawn in an arc, circle = radial lines drawn from a central point etc.)

Basic idea is we determine the maximum number of grids that can be affected, then from the origin point, iterate along the width of the shape to cast forward to the maximum length. Anything in the resultant grid set will be affected by the ability. If we run out of width, or exceed the maximum aoe of the ability, that ends things.

Works when cast in specific directions but not when target is not in a direct cardinal direction from origin
Need to refactor to account for target co-ordinates, wil probably need some vector math (numpy?)
Branching this commit as it breaks the semi-functional breath attack that was at least partially working